weibo_pipeline/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
openai>=1.12.0
urllib3>=2.0.0
